requirements.txt
numpy
pandas
scikit-learn
nltk
rank_bm25
faiss-cpu
joblib
tqdm
